waitN: Fix index bug when there are handles before UEvents in the array

+++ nx/source/kernel/uevent.h
@@ -3,4 +3,4 @@
 #include "kernel/uevent.h"
 
 void _ueventTryAutoClear(UEvent* e);
-bool _ueventAddListener(UEvent* e, WaiterNode* w, size_t idx, size_t* idx_out, Handle thread);
+bool _ueventAddListener(UEvent* e, WaiterNode* w, s32 idx, s32* idx_out, Handle thread);
